Incorporated the Pattern class into the matcher module.  Deleted the pattern module.
