+++ stix2matcher/matcher.py
@@ -21,7 +21,7 @@ import antlr4
 import antlr4.error.Errors
 import antlr4.error.ErrorListener
 
-import stix2matcher.pattern
+import stix2patterns.pattern
 from stix2patterns.grammars.STIXPatternListener import STIXPatternListener
 from stix2patterns.grammars.STIXPatternParser import STIXPatternParser
 
@@ -2062,6 +2062,46 @@ class MatchListener(STIXPatternListener):
         self.__push(s, u"exitSetLiteral ({})".format(ctx.getText()))
 
 
+class Pattern(stix2patterns.pattern.Pattern):
+    """
+    Represents a pattern in a "compiled" form, for more efficient reuse.
+    """
+
+    def __init__(self, pattern_str):
+        """
+        Compile a pattern.
+
+        :param pattern_str: The pattern to compile
+        :raises stix2patterns.pattern.ParseException: If there is a parse error
+        """
+        super(Pattern, self).__init__(pattern_str)
+
+    def match(self, observed_data_sdos, verbose=False):
+        """
+        Match this pattern against the given observations.  Returns matching
+        SDOs.  The matcher can find many bindings; this function returns the
+        SDOs corresponding to only the first binding found.
+
+        :param observed_data_sdos: A list of observed-data SDOs, as a list of
+            dicts.  STIX JSON should be parsed into native Python structures
+            before calling this function.
+        :param verbose: Whether to dump detailed info about matcher operation
+        :return: Matching SDOs if the pattern matched; an empty list if it
+            didn't match.
+        :raises MatcherException: If an error occurs during matching
+        """
+        matcher = MatchListener(observed_data_sdos, verbose)
+        antlr4.ParseTreeWalker.DEFAULT.walk(matcher, self.__parse_tree)
+
+        found_bindings = matcher.matched()
+        if found_bindings:
+            matching_sdos = matcher.get_sdos_from_binding(found_bindings[0])
+        else:
+            matching_sdos = []
+
+        return matching_sdos
+
+
 def match(pattern, observed_data_sdos, verbose=False):
     """
     Match the given pattern against the given observations.  Returns matching
@@ -2075,9 +2115,11 @@ def match(pattern, observed_data_sdos, verbose=False):
     :param verbose: Whether to dump detailed info about matcher operation
     :return: Matching SDOs if the pattern matched; an empty list if it didn't
         match.
+    :raises stix2patterns.pattern.ParseException: If there is a parse error
+    :raises MatcherException: If an error occurs during matching
     """
 
-    compiled_pattern = stix2matcher.pattern.Pattern(pattern)
+    compiled_pattern = Pattern(pattern)
     return compiled_pattern.match(observed_data_sdos, verbose)
